Move AsPolygonSet out of package file.
Moving AsPolygonSet out of the package file allows us to dodge a
compiler error.
